requirements.txt
fastapi==0.115.12
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
together==0.2.4
schedule==1.2.1
nltk==3.8.1
transformers>=4.41.0,<5.0.0
torch==2.2.2
scikit-learn==1.3.2
openai==1.70.0
httpx==0.28.1
google-genai==1.11.0
sentence-transformers>=4.0.2
huggingface-hub>=0.30.1
strsimpy==0.2.1
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
spacy>=3.0.0
feedparser>=6.0.0
beautifulsoup4>=4.9.1
dateparser>=1.1.0
requests>=2.31.0
